requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
openai==2.8.1
google-genai==1.52.0
tiktoken==0.12.0